pyproject.toml
[project]
name = "evolaris-academy"
version = "0.1.0"
description = "LMS Professional con Django 6"
authors = [
    {name = "Daniel Granada"}
]
license = {text = "MIT"}
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "pillow (>=12.1.0,<13.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.0,<0.16.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "django-debug-toolbar (>=6.2.0,<7.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "U"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"config/settings/*.py" = ["F403", "F405"]
"**/migrations/*" = ["ALL"]